Prueba

+++ app/src/main/java/es/elconfidencial/eleccionesec/chart/PieChartRendererEC.java
@@ -136,7 +136,7 @@ public class PieChartRendererEC extends PieChartRenderer {
 
                 //Ocultamos los % muy pequeños
                 String val = dataSet.getValueFormatter().getFormattedValue(value);
-                if (Integer.parseInt(val.split(",")[0]) <= 1){
+                if (Double.parseDouble(val.split(",")[0]) <= 1){
                     val = "";
                 }
